split llk files
